requirements.txt
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.1
django-filter==2.1.0
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.9.4
gunicorn==19.9.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.0
numpy==1.16.4
psycopg2==2.8.3
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
response==0.3.0
scipy==1.3.0
selenium==3.141.0
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
whitenoise==4.1.2
